Split cxl ide test cases into single file
