--- teeio-validator/library/cxl_ide_test_lib/test_case/test_case_key_prog.c
@@ -1,173 +1,0 @@
-/**
- *  Copyright Notice:
- *  Copyright 2023-2024 Intel. All rights reserved.
- *  License: BSD 3-Clause License.
- **/
-
-#include <stdlib.h>
-#include <stdio.h>
-#include <ctype.h>
-#include "ide_test.h"
-#include "teeio_debug.h"
-
-bool cxl_ide_test_key_prog_1_setup(void *test_context)
-{
-  NOT_IMPLEMENTED(__func__);
-  return false;
-}
-
-bool cxl_ide_test_key_prog_1_run(void *test_context)
-{
-  NOT_IMPLEMENTED(__func__);
-  return false;
-}
-
-bool cxl_ide_test_key_prog_1_teardown(void *test_context)
-{
-  NOT_IMPLEMENTED(__func__);
-  return false;
-}
-
-bool cxl_ide_test_key_prog_2_setup(void *test_context)
-{
-  NOT_IMPLEMENTED(__func__);
-  return false;
-}
-
-bool cxl_ide_test_key_prog_2_run(void *test_context)
-{
-  NOT_IMPLEMENTED(__func__);
-  return false;
-}
-
-bool cxl_ide_test_key_prog_2_teardown(void *test_context)
-{
-  NOT_IMPLEMENTED(__func__);
-  return false;
-}
-
-bool cxl_ide_test_key_prog_3_setup(void *test_context)
-{
-  NOT_IMPLEMENTED(__func__);
-  return false;
-}
-
-bool cxl_ide_test_key_prog_3_run(void *test_context)
-{
-  NOT_IMPLEMENTED(__func__);
-  return false;
-}
-
-bool cxl_ide_test_key_prog_3_teardown(void *test_context)
-{
-  NOT_IMPLEMENTED(__func__);
-  return false;
-}
-
-bool cxl_ide_test_key_prog_4_setup(void *test_context)
-{
-  NOT_IMPLEMENTED(__func__);
-  return false;
-}
-
-bool cxl_ide_test_key_prog_4_run(void *test_context)
-{
-  NOT_IMPLEMENTED(__func__);
-  return false;
-}
-
-bool cxl_ide_test_key_prog_4_teardown(void *test_context)
-{
-  NOT_IMPLEMENTED(__func__);
-  return false;
-}
-
-bool cxl_ide_test_key_prog_5_setup(void *test_context)
-{
-  NOT_IMPLEMENTED(__func__);
-  return false;
-}
-
-bool cxl_ide_test_key_prog_5_run(void *test_context)
-{
-  NOT_IMPLEMENTED(__func__);
-  return false;
-}
-
-bool cxl_ide_test_key_prog_5_teardown(void *test_context)
-{
-  NOT_IMPLEMENTED(__func__);
-  return false;
-}
-
-bool cxl_ide_test_key_prog_6_setup(void *test_context)
-{
-  NOT_IMPLEMENTED(__func__);
-  return false;
-}
-
-bool cxl_ide_test_key_prog_6_run(void *test_context)
-{
-  NOT_IMPLEMENTED(__func__);
-  return false;
-}
-
-bool cxl_ide_test_key_prog_6_teardown(void *test_context)
-{
-  NOT_IMPLEMENTED(__func__);
-  return false;
-}
-
-bool cxl_ide_test_key_prog_7_setup(void *test_context)
-{
-  NOT_IMPLEMENTED(__func__);
-  return false;
-}
-
-bool cxl_ide_test_key_prog_7_run(void *test_context)
-{
-  NOT_IMPLEMENTED(__func__);
-  return false;
-}
-
-bool cxl_ide_test_key_prog_7_teardown(void *test_context)
-{
-  NOT_IMPLEMENTED(__func__);
-  return false;
-}
-
-bool cxl_ide_test_key_prog_8_setup(void *test_context)
-{
-  NOT_IMPLEMENTED(__func__);
-  return false;
-}
-
-bool cxl_ide_test_key_prog_8_run(void *test_context)
-{
-  NOT_IMPLEMENTED(__func__);
-  return false;
-}
-
-bool cxl_ide_test_key_prog_8_teardown(void *test_context)
-{
-  NOT_IMPLEMENTED(__func__);
-  return false;
-}
-
-bool cxl_ide_test_key_prog_9_setup(void *test_context)
-{
-  NOT_IMPLEMENTED(__func__);
-  return false;
-}
-
-bool cxl_ide_test_key_prog_9_run(void *test_context)
-{
-  NOT_IMPLEMENTED(__func__);
-  return false;
-}
-
-bool cxl_ide_test_key_prog_9_teardown(void *test_context)
-{
-  NOT_IMPLEMENTED(__func__);
-  return false;
-}
