+++ teeio-validator/library/cxl_ide_test_lib/test_case/test_case_key_prog_6.c
@@ -1,0 +1,29 @@
+/**
+ *  Copyright Notice:
+ *  Copyright 2023-2024 Intel. All rights reserved.
+ *  License: BSD 3-Clause License.
+ **/
+
+#include <stdlib.h>
+#include <stdio.h>
+#include <ctype.h>
+#include "ide_test.h"
+#include "teeio_debug.h"
+
+bool cxl_ide_test_key_prog_6_setup(void *test_context)
+{
+  NOT_IMPLEMENTED(__func__);
+  return false;
+}
+
+bool cxl_ide_test_key_prog_6_run(void *test_context)
+{
+  NOT_IMPLEMENTED(__func__);
+  return false;
+}
+
+bool cxl_ide_test_key_prog_6_teardown(void *test_context)
+{
+  NOT_IMPLEMENTED(__func__);
+  return false;
+}
